Rename /api to /core
