+++ example/pages/_api.code-search
@@ -1,0 +1,94 @@
+# Query: /api
+# ContextLines: 1
+
+26 résultats - Fichiers 13
+
+README.md:
+  392    History mode can
+  393:   be [BROWSER](https://github.com/ReactTraining/history/blob/master/docs/api-reference.md#createbrowserhistory)
+  394    ,
+  395:   [HASH](https://github.com/ReactTraining/history/blob/master/docs/api-reference.md#createhashhistory)
+  396    ,
+  397:   [MEMORY](https://github.com/ReactTraining/history/blob/master/docs/api-reference.md#creatememoryhistory)
+  398    . For more information, check
+  399:   the [history library documentation](https://github.com/ReactTraining/history/blob/master/docs/api-reference.md)
+  400  - **middlewares** `[]` add routes middleware function to patch each routes (check [langMiddleware](src/lang/LangMiddleware.ts) example)
+
+example/pages/AboutPage.tsx:
+  10    getSubRouterRoutes,
+  11: } from "../../src/core/helpers";
+  12  import { routesList } from "../routes";
+
+example/pages/HelloPage.tsx:
+  8    joinPaths,
+  9: } from "../../src/core/helpers";
+  10  import { transitionsHelper } from "../helper/transitionsHelper";
+
+src/index.ts:
+   1: export { Routers } from "./core/Routers";
+   2  export { Router } from "./components/Router";
+
+   9    getRoutePathByRouteName,
+  10: } from "./core/helpers";
+  11: export type { TOpenRouteParams } from "./core/helpers";
+  12  
+
+  22  
+  23: export { default as LangService } from "./core/LangService";
+  24: export type { TLanguage } from "./core/LangService";
+  25  
+
+src/components/Link.tsx:
+  7    TOpenRouteParams,
+  8: } from "../core/helpers";
+  9  import debug from "@wbe/debug";
+
+src/components/Router.tsx:
+   9  import React from "react";
+  10: import { applyMiddlewares, patchMissingRootRoute } from "../core/helpers";
+  11: import { getNotFoundRoute, getRouteFromUrl } from "../core/matcher";
+  12: import { Routers } from "../core/Routers";
+  13: import LangService from "../core/LangService";
+  14  
+
+src/core/LangService.ts:
+   1: import { Routers } from "../core/Routers";
+   2  import {
+
+   6    removeLastCharFromString,
+   7: } from "../core/helpers";
+   8  import debug from "@wbe/debug";
+   9  import { TRoute } from "../components/Router";
+  10: import { getLangPathByLang } from "../core/helpers";
+  11  
+
+src/core/Routers.ts:
+  1: import LangService from "../core/LangService";
+  2  import { BrowserHistory, HashHistory, MemoryHistory } from "history";
+
+src/hooks/useLocation.ts:
+  2  import { useHistory } from "../hooks/useHistory";
+  3: import { createUrl, TOpenRouteParams } from "../core/helpers";
+  4  import debug from "@wbe/debug";
+
+src/hooks/useRouteCounter.ts:
+  2  import { useState } from "react";
+  3: import { Routers } from "../core/Routers";
+  4  
+
+test/helpers.test.ts:
+  8    removeLastCharFromString,
+  9: } from "../src/core/helpers";
+  10  import { TRoute } from "../src/components/Router";
+
+test/Link.test.tsx:
+  4  import { render, fireEvent } from "@testing-library/react";
+  5: import { Routers } from "../src/core/Routers";
+  6  import { LangService } from "../src";
+  7: import { TOpenRouteParams } from "../src/core/helpers";
+  8  import { TRoute } from "../src/components/Router";
+
+test/matcher.test.ts:
+  1  // import { Routers, RouterManager, TRoute } from "../src";
+  2: // import { preventSlashes } from "../src/core/helpers";
+  3
