+++ src/hooks/useLocation.ts
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useHistory } from "../hooks/useHistory";
-import { createUrl, TOpenRouteParams } from "../api/helpers";
+import { createUrl, TOpenRouteParams } from "../core/helpers";
 import debug from "@wbe/debug";
 const log = debug("router:useLocation");
