+++ src/index.ts
@@ -1,4 +1,4 @@
-export { Routers } from "./api/Routers";
+export { Routers } from "./core/Routers";
 export { Router } from "./components/Router";
 export { Link } from "./components/Link";
 export {
@@ -7,8 +7,8 @@ export {
   getSubRouterBase,
   getSubRouterRoutes,
   getRoutePathByRouteName,
-} from "./api/helpers";
-export type { TOpenRouteParams } from "./api/helpers";
+} from "./core/helpers";
+export type { TOpenRouteParams } from "./core/helpers";
 
 export { Stack } from "./components/Stack";
 export type { TManageTransitions } from "./components/Stack";
@@ -20,5 +20,5 @@ export { useRouteCounter } from "./hooks/useRouteCounter";
 export { useStack } from "./hooks/useStack";
 export type { IRouteStack } from "./hooks/useStack";
 
-export { default as LangService } from "./api/LangService";
-export type { TLanguage } from "./api/LangService";
+export { default as LangService } from "./core/LangService";
+export type { TLanguage } from "./core/LangService";
